+++ proc.c
@@ -61,7 +61,7 @@ pic_proc_apply(pic_state *pic)
 {
   struct pic_proc *proc;
   pic_value *args;
-  int argc;
+  size_t argc;
   pic_value arg_list;
 
   pic_get_args(pic, "l*", &proc, &argc, &args);
